+++ src/shared/utils/index.ts
@@ -1,5 +1,7 @@
 import * as prettier from "prettier";
 
+import { nextTick } from "node:process";
+
 import { CommonUtils } from "@utils/common";
 
 import { specialWordsMap } from "../init";
@@ -73,19 +75,19 @@ export function reIndent(str: string, indentNum = 0) {
         .join("\n");
 }
 
-export function prettierFormatFile(filePath: string) {
-    process.nextTick(async () => {
-        const originalData = await fs.promises.readFile(filePath, {
-            encoding: "utf8",
-        });
+export async function prettierFormatFile(filePath: string) {
+    return await new Promise((resolve: (text: string) => void) => {
+        nextTick(async () => {
+            const originalData = await fs.promises.readFile(filePath, {
+                encoding: "utf8",
+            });
 
-        const formattedData = prettier.format(originalData, {
-            parser: "typescript",
-            tabWidth: 4,
-        });
-
-        await fs.promises.writeFile(filePath, formattedData, {
-            encoding: "utf8",
+            resolve(
+                prettier.format(originalData, {
+                    parser: "typescript",
+                    tabWidth: 4,
+                })
+            );
         });
     });
 }